+++ zatgo_core/hooks.py
@@ -13,7 +13,7 @@ app_description = (
 )
 app_email = "engineering@zatgo.local"
 app_license = "mit"
-app_version = "0.2.1"
+app_version = "0.2.2"
 
 required_apps = ["erpnext"]